Testing location implemented.
* Duplicated the Player/MovingBody so that we can have the exact same format but for testing.
* Implemented a straight testing path with a few segments
* Duplicated the Moving-Body with the path all setup

To run Test 1, 2 and 3 you change the settings for the specific Test Body and disable all the others in order to run with the Test-config.

If more tests are needed, copy the Baseline Test Moving-Body and start the config from scratch with all the elements set to 0 as a base.